+++ lib/core/src/string_extension.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:morpheme_cli/core/src/commandline_converter.dart';
+import 'package:morpheme_cli/helper/helper.dart';
 
 extension StringExtension on String {
   Future<int> get run async {
@@ -18,7 +19,15 @@ extension StringExtension on String {
     );
     process.stdout.transform(utf8.decoder).listen(stdout.write);
     process.stderr.transform(utf8.decoder).listen(stderr.write);
-    return process.exitCode;
+
+    final exitCode = await process.exitCode;
+
+    if (exitCode > 0) {
+      StatusHelper.failed('$this has exit with code $exitCode',
+          statusExit: exitCode);
+    }
+
+    return exitCode;
   }
 
   Future<int> start({
@@ -50,7 +59,13 @@ extension StringExtension on String {
       if (showLog) stderr.write(line);
     }));
 
-    return process.exitCode;
+    final exitCode = await process.exitCode;
+
+    if (exitCode > 0) {
+      StatusHelper.failed('$this has exit with code $exitCode',
+          statusExit: exitCode);
+    }
+    return exitCode;
   }
 
   void write(String line) =>